Move git helper to own package
